Разбиение на пакеты Рефакторинг кода Добавлен DEVELOPMENT.md Обновлён README.md
