+++ DEVELOPMENT.md
@@ -1,0 +1,48 @@
+# Процесс разработки
+## Task 1
+******************
+Данное задание заключается в создании _консольного бота_, который просто пересылает полученное сообщение.
+Необходимо реализовать заданный функционал с использованием интерфейсов `Reader`, `Writer`, `Handler`.
+
+Статус - <span style="color:green">**Зачтено**</span>
+
+## Task 2
+********************
+Необходимо реализовать работу _консольного_ и _Telegram_ бота с использованием общих обработчиков.
+
+- Выполнено разбиение на пакеты
+- Добавлена возможность лёгкого добавления новых обработчиков
+- Каждый бот имеет свой собственный файл запуска. При запуске через главный `Main` файл будет предложен выбор бота
+  для запуска
+
+Статус - <span style="color:green">**Зачтено**</span>
+
+## Task 3
+*********************
+- Необходимо реализовать отправку HTTP запросов и получение расписания с сайта УрФУ.
+- Необходимо выполнить преобразование полученных данных в сущности кабинетов со всей необходимой информацией.
+
+Статус - <span style="color:orange">**В разработке**</span>
+
+## Task 4
+*********************
+- Необходимо реализовать подключение к базе данных
+- Необходимо организовать запись данных о кабинетах в БД
+- Сохранять информацию о пользователях и их выборе в БД
+- Реализовать запрос данных с сайта по определённому времени
+
+Статус - <span style="color:red">**Ожидает подтверждения**</span>
+
+## Task 5
+*********************
+- Добавить регистрацию в боте
+- Добавить возможность просмотра свободных аудиторий
+
+Статус - <span style="color:red">**Ожидает подтверждения**</span>
+
+## Task 6
+*********************
+- Добавить возможность исправления статуса аудитории
+- Провести финальный рефакторинг
+
+Статус - <span style="color:red">**Ожидает подтверждения**</span>
