+++ README.md
@@ -1,5 +1,10 @@
 # JavaTelgramBot
-Проект по ООП 3 семестр. Бот для телеграм
+>Проект по ООП 3 семестр. Бот для телеграм 
+
+>Информация о [процессе разработки](DEVELOPMENT.md)  
 
 Данный бот позволяет определять свободные аудитории в различных институтах, 
-основываясь на данных сайта УрФУ и поправках от пользователей
+основываясь на данных сайта УрФУ и поправках от пользователей.  
+
+### Доступные команды
+<span style="color:gray">_В разработке_</span>
