+++ src/main/java/ru/telegramBot/gm/Main.java
@@ -1,78 +1,56 @@
 package ru.telegramBot.gm;
 
 
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-import ru.telegramBot.gm.dataComponents.TextComponent;
-import ru.telegramBot.gm.handlers.HandlerFacade;
-import ru.telegramBot.gm.handlers.Handler;
-import ru.telegramBot.gm.consoleBot.ConsoleReader;
-import ru.telegramBot.gm.readers.Reader;
-import ru.telegramBot.gm.readers.RequestData;
-import ru.telegramBot.gm.telegramBot.TelegramBot;
-import ru.telegramBot.gm.consoleBot.ConsoleWriter;
-import ru.telegramBot.gm.writers.ResponseData;
-import ru.telegramBot.gm.writers.Writer;
-
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.util.Objects;
-import java.util.Properties;
+import java.util.Scanner;
 
 public class Main {
-
-    public static void main(String[] args) throws IOException {
-        telegramBot();
-//        consoleEchoBot();
-    }
-
     /**
-     * Простой эхо ответчик для консоли.
-     * Останавливает считывание и ответ при вводе пустой строки
-     */
-    public static void consoleEchoBot() {
-        Reader reader = new ConsoleReader();
-        Writer writer = new ConsoleWriter();
-        Handler handler = new HandlerFacade();
-
-        while (true) {
-            RequestData requestData = reader.read();
-            TextComponent textComponent = requestData.getComponent(TextComponent.class);
-            if (textComponent == null)
-                break;
-            if (Objects.equals(textComponent.getText(), ""))
-                break;
-            ResponseData responseData = handler.handle(requestData);
-            writer.write(responseData);
-        }
-    }
-
-    /**
-     * Подключение и запуск сессии телеграм. Данные бота из конфигурационного файла
-     */
-    public static void telegramBot() throws IOException {
-        Properties properties = new Properties();
-        properties.load(new FileInputStream("src/main/resources/config.properties"));
-
-        String botName = Objects.requireNonNull(properties.getProperty("bot.name"), "Не найдено bot.name в файле конфигурации");
-        String botToken = Objects.requireNonNull(properties.getProperty("bot.token"), "Не найдено bot.token в файле конфигурации");
-
-        startTelegramBot(botName, botToken);
-    }
-
-    /**
-     * Данный метод создаёт и регистрирует телеграм бота
+     * Главный метод запуска бота, который предоставляет выбор, какого бота запускать
      *
-     * @param botName Имя бота
-     * @param botToken Токен, для подключения к боту
+     * @param args Не используются
      */
-    private static void startTelegramBot(String botName, String botToken){
-        try {
-            TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
-            botsApi.registerBot(new TelegramBot(botName, botToken));
-        } catch (TelegramApiException e) {
-            e.printStackTrace();
+    public static void main(String[] args) {
+        int result = chooseBot();
+        switch (result) {
+            case 0:
+                break;
+            case 1:
+                ru.telegramBot.gm.clients.telegramBot.Main.main(new String[]{});
+                break;
+            case 2:
+                ru.telegramBot.gm.clients.consoleBot.Main.main(new String[]{});
         }
+    }
+
+    /**
+     * Метод, который получает у пользователя информацию о том, какого бота необходимо запустить
+     *
+     * @return <ul><li>0 - exit</li><li>1 - Telegram bot</li><li>2 - Console bot</li></ul>
+     */
+    private static int chooseBot() {
+        System.out.print("Choose which bot you want to run (t/c/e/?): ");
+        Scanner scanner = new Scanner(System.in);
+        String answer = scanner.next();
+        return switch (answer) {
+            case "e" -> {
+                System.out.println("Bot start canceled");
+                yield 0;
+            }
+            case "t" -> {
+                System.out.println("Starting a telegram bot...");
+                yield 1;
+            }
+            case "c" -> {
+                System.out.println("Starting a console bot...");
+                yield 2;
+            }
+            default -> {
+                System.out.println("t - Starting a telegram bot. The connection data is obtained from the config file");
+                System.out.println("c - Starting a console bot");
+                System.out.println("e - (exit) cancel the start of the bot");
+                System.out.println("? - getting help on launching bots");
+                yield chooseBot();
+            }
+        };
     }
 }
