+++ src/main/java/ru/telegramBot/gm/app/dataComponents/Component.java
@@ -1,0 +1,10 @@
+package ru.telegramBot.gm.app.dataComponents;
+
+/**
+ * Класс компонента контейнера данных
+ */
+public interface Component {
+    int hashCode();
+
+    boolean equals(final Object obj);
+}
