--- src/main/java/ru/telegramBot/gm/dataComponents/Component.java
@@ -1,9 +1,0 @@
-package ru.telegramBot.gm.dataComponents;
-
-/**
- * Класс компонента контейнера данных
- */
-public interface Component {
-    int hashCode();
-    boolean equals(final Object obj);
-}
