--- src/main/java/ru/telegramBot/gm/handlers/textHandlers/TextFacade.java
@@ -1,16 +1,0 @@
-package ru.telegramBot.gm.handlers.textHandlers;
-
-import ru.telegramBot.gm.handlers.BaseHandlersFacade;
-
-/**
- * Фасад, который позволяет обрабатывать тексты, не содержащие команды
- */
-public class TextFacade extends BaseHandlersFacade {
-
-    /**
-     * Задание списка всех обработчиков, которым даётся возможность обработать данные
-     */
-    public TextFacade(){
-        super(new SimpleEchoHandler());
-    }
-}
