--- src/main/java/ru/telegramBot/gm/readers/Reader.java
@@ -1,13 +1,0 @@
-package ru.telegramBot.gm.readers;
-
-/**
- * Интерфейс обеспечивающий чтение данных из разных источников
- * Всегда возвращает упакованный контейнер
- */
-public interface Reader {
-    /**
-     * Метод, используемый для чтения данных из какого-либо источника
-     * @return Упакованный контейнер с прочтёнными данными
-     */
-    RequestData read();
-}
