--- src/main/java/ru/telegramBot/gm/writers/ResponseData.java
@@ -1,17 +1,0 @@
-package ru.telegramBot.gm.writers;
-
-import ru.telegramBot.gm.dataComponents.ComponentsContainer;
-
-
-
-/**
- * Контейнер для обработанных данных, который содержит компоненты с данными. Обращение к компонентам по имени
- */
-public class ResponseData extends ComponentsContainer {
-    public ResponseData() {
-        super();
-    }
-    public ResponseData(ResponseData oldData){
-        super(oldData);
-    }
-}
